DIG-24 * move screen specific components into respective screen directories
